+++ app.py
@@ -48,7 +48,7 @@ from utils import get_settings, utc_now
 from routes import register_all
 from messages import TRANSLATIONS
 
-APP_VERSION = '1.72.4'
+APP_VERSION = '1.72.5'
 
 csrf = CSRFProtect()